Refactor type_traits, forward and concepts
