+++ include/boost/safe_numbers/bit.hpp
@@ -6,7 +6,6 @@
 #define BOOST_SAFE_NUMBERS_BIT_HPP
 
 #include <boost/safe_numbers/detail/unsigned_integer_basis.hpp>
-#include <boost/safe_numbers/detail/concepts.hpp>
 #include <boost/safe_numbers/detail/type_traits.hpp>
 #include <boost/safe_numbers/detail/int128/bit.hpp>
